requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.12
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
oauth2client==3.0.0
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tornado==6.1
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.8
